EES-5492 - broke orchestrations into their own simplified files and made static, to avoid issues whereby exceptions thrown during the construction of Function classes caused the ImportMetadata stage to lock up
