+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Tests/Functions/ProcessCompletionOfNextDataSetVersionImportFunctionTests.cs
@@ -7,14 +7,9 @@ using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.Parquet.Tables;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.Tests.Fixtures;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Functions;
-using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Model;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Utils;
-using Microsoft.DurableTask;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging.Abstractions;
-using Moq;
-using static GovUk.Education.ExploreEducationStatistics.Common.Tests.Utils.MockUtils;
 using FilterMeta = GovUk.Education.ExploreEducationStatistics.Public.Data.Model.FilterMeta;
 
 namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Tests.Functions;
@@ -37,93 +32,6 @@ public abstract class ProcessCompletionOfNextDataSetVersionImportFunctionTests(
         TimePeriodsTable.ParquetFile
     ];
 
-    public class ProcessCompletionOfNextDataSetVersionImportTests(
-        ProcessorFunctionsIntegrationTestFixture fixture)
-        : ProcessCompletionOfNextDataSetVersionImportFunctionTests(fixture)
-    {
-        [Fact]
-        public async Task Success()
-        {
-            var mockOrchestrationContext = DefaultMockOrchestrationContext();
-            var activitySequence = new MockSequence();
-
-            string[] expectedActivitySequence =
-            [
-                ActivityNames.UpdateFileStoragePath,
-                ActivityNames.ImportMetadata,
-                ActivityNames.CreateChanges,
-                ActivityNames.ImportData,
-                ActivityNames.WriteDataFiles,
-                ActivityNames.CompleteNextDataSetVersionImportProcessing
-            ];
-
-            foreach (var activityName in expectedActivitySequence)
-            {
-                mockOrchestrationContext
-                    .InSequence(activitySequence)
-                    .Setup(context => context.CallActivityAsync(activityName,
-                        mockOrchestrationContext.Object.InstanceId,
-                        null))
-                    .Returns(Task.CompletedTask);
-            }
-
-            await ProcessCompletionOfNextDataSetVersionImport(mockOrchestrationContext.Object);
-
-            VerifyAllMocks(mockOrchestrationContext);
-        }
-
-        [Fact]
-        public async Task ActivityFunctionThrowsException_CallsHandleFailureActivity()
-        {
-            var mockOrchestrationContext = DefaultMockOrchestrationContext();
-
-            var activitySequence = new MockSequence();
-
-            mockOrchestrationContext
-                .InSequence(activitySequence)
-                .Setup(context =>
-                    context.CallActivityAsync(ActivityNames.UpdateFileStoragePath,
-                        mockOrchestrationContext.Object.InstanceId,
-                        null))
-                .Throws<Exception>();
-
-            mockOrchestrationContext
-                .InSequence(activitySequence)
-                .Setup(context =>
-                    context.CallActivityAsync(ActivityNames.HandleProcessingFailure,
-                        mockOrchestrationContext.Object.InstanceId,
-                        null))
-                .Returns(Task.CompletedTask);
-
-            await ProcessCompletionOfNextDataSetVersionImport(mockOrchestrationContext.Object);
-
-            VerifyAllMocks(mockOrchestrationContext);
-        }
-
-        private async Task ProcessCompletionOfNextDataSetVersionImport(TaskOrchestrationContext orchestrationContext)
-        {
-            var function = GetRequiredService<ProcessCompletionOfNextDataSetVersionFunction>();
-            await function.ProcessCompletionOfNextDataSetVersion(
-                orchestrationContext,
-                new ProcessDataSetVersionContext { DataSetVersionId = Guid.NewGuid() });
-        }
-
-        private static Mock<TaskOrchestrationContext> DefaultMockOrchestrationContext(Guid? instanceId = null)
-        {
-            var mock = new Mock<TaskOrchestrationContext>(MockBehavior.Strict);
-
-            mock.Setup(context =>
-                    context.CreateReplaySafeLogger(
-                        nameof(ProcessCompletionOfNextDataSetVersionFunction.ProcessCompletionOfNextDataSetVersion)))
-                .Returns(NullLogger.Instance);
-
-            mock.SetupGet(context => context.InstanceId)
-                .Returns(instanceId?.ToString() ?? Guid.NewGuid().ToString());
-
-            return mock;
-        }
-    }
-
     public abstract class CreateChangesTests(
         ProcessorFunctionsIntegrationTestFixture fixture)
         : ProcessCompletionOfNextDataSetVersionImportFunctionTests(fixture)
@@ -132,7 +40,7 @@ public abstract class ProcessCompletionOfNextDataSetVersionImportFunctionTests(
 
         protected async Task CreateChanges(Guid instanceId)
         {
-            var function = GetRequiredService<ProcessCompletionOfNextDataSetVersionFunction>();
+            var function = GetRequiredService<ProcessCompletionOfNextDataSetVersionFunctions>();
             await function.CreateChanges(instanceId, CancellationToken.None);
         }
     }
@@ -3132,7 +3040,7 @@ public abstract class ProcessCompletionOfNextDataSetVersionImportFunctionTests(
 
         private async Task UpdateFileStoragePath(Guid instanceId)
         {
-            var function = GetRequiredService<ProcessCompletionOfNextDataSetVersionFunction>();
+            var function = GetRequiredService<ProcessCompletionOfNextDataSetVersionFunctions>();
             await function.UpdateFileStoragePath(instanceId, CancellationToken.None);
         }
     }
@@ -3190,7 +3098,7 @@ public abstract class ProcessCompletionOfNextDataSetVersionImportFunctionTests(
 
         private async Task CompleteProcessing(Guid instanceId)
         {
-            var function = GetRequiredService<ProcessCompletionOfNextDataSetVersionFunction>();
+            var function = GetRequiredService<ProcessCompletionOfNextDataSetVersionFunctions>();
             await function.CompleteNextDataSetVersionImportProcessing(instanceId, CancellationToken.None);
         }
     }
