+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Functions/ProcessInitialDataSetVersionOrchestration.cs
@@ -1,20 +1,15 @@
-using GovUk.Education.ExploreEducationStatistics.Public.Data.Model;
-using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Extensions;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Model;
-using GovUk.Education.ExploreEducationStatistics.Public.Data.Services.Interfaces;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.DurableTask;
 using Microsoft.Extensions.Logging;
 
 namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Functions;
 
-public class ProcessInitialDataSetVersionFunction(
-    PublicDataDbContext publicDataDbContext,
-    IDataSetVersionPathResolver dataSetVersionPathResolver) : BaseProcessDataSetVersionFunction(publicDataDbContext)
+public static class ProcessInitialDataSetVersionOrchestration
 {
     [Function(nameof(ProcessInitialDataSetVersion))]
-    public async Task ProcessInitialDataSetVersion(
+    public static async Task ProcessInitialDataSetVersion(
         [OrchestrationTrigger] TaskOrchestrationContext context,
         ProcessDataSetVersionContext input)
     {
@@ -43,24 +38,5 @@ public class ProcessInitialDataSetVersionFunction(
 
             await context.CallActivity(ActivityNames.HandleProcessingFailure, logger, context.InstanceId);
         }
-    }
-
-    [Function(ActivityNames.CompleteInitialDataSetVersionProcessing)]
-    public async Task CompleteInitialDataSetVersionProcessing(
-        [ActivityTrigger] Guid instanceId,
-        CancellationToken cancellationToken)
-    {
-        var dataSetVersionImport = await GetDataSetVersionImport(instanceId, cancellationToken);
-        await UpdateImportStage(dataSetVersionImport, DataSetVersionImportStage.Completing, cancellationToken);
-
-        var dataSetVersion = dataSetVersionImport.DataSetVersion;
-
-        // Delete the DuckDb database file as it is no longer needed
-        File.Delete(dataSetVersionPathResolver.DuckDbPath(dataSetVersion));
-
-        dataSetVersion.Status = DataSetVersionStatus.Draft;
-
-        dataSetVersionImport.Completed = DateTimeOffset.UtcNow;
-        await publicDataDbContext.SaveChangesAsync(cancellationToken);
     }
 }
