+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Functions/ProcessNextDataSetVersionMappingsFunctions.cs
@@ -1,48 +1,14 @@
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Model;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.Database;
-using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Extensions;
-using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Model;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Services.Interfaces;
 using Microsoft.Azure.Functions.Worker;
-using Microsoft.DurableTask;
-using Microsoft.Extensions.Logging;
 
 namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Functions;
 
-public class ProcessNextDataSetVersionMappingsFunction(
+public class ProcessNextDataSetVersionMappingsFunctions(
     PublicDataDbContext publicDataDbContext,
     IDataSetVersionMappingService mappingService) : BaseProcessDataSetVersionFunction(publicDataDbContext)
 {
-    [Function(nameof(ProcessNextDataSetVersionMappings))]
-    public async Task ProcessNextDataSetVersionMappings([OrchestrationTrigger] TaskOrchestrationContext context,
-        ProcessDataSetVersionContext input)
-    {
-        var logger = context.CreateReplaySafeLogger(nameof(ProcessNextDataSetVersionMappings));
-
-        logger.LogInformation(
-            "Processing next data set version (InstanceId={InstanceId}, DataSetVersionId={DataSetVersionId})",
-            context.InstanceId,
-            input.DataSetVersionId);
-
-        try
-        {
-            await context.CallActivity(ActivityNames.CopyCsvFiles, logger, context.InstanceId);
-            await context.CallActivity(ActivityNames.CreateMappings, logger, context.InstanceId);
-            await context.CallActivity(ActivityNames.ApplyAutoMappings, logger, context.InstanceId);
-            await context.CallActivity(ActivityNames.CompleteNextDataSetVersionMappingProcessing, logger,
-                context.InstanceId);
-        }
-        catch (Exception e)
-        {
-            logger.LogError(e,
-                "Activity failed with an exception (InstanceId={InstanceId}, DataSetVersionId={DataSetVersionId})",
-                context.InstanceId,
-                input.DataSetVersionId);
-
-            await context.CallActivity(ActivityNames.HandleProcessingFailure, logger, context.InstanceId);
-        }
-    }
-
     [Function(ActivityNames.CreateMappings)]
     public async Task CreateMappings(
         [ActivityTrigger] Guid instanceId,
